Adds Larger Size for Small Screens, Disable/Enable, Style Updates
Adds a media query to make the rating steps larger on small screens
to make the click areas easier to hit on mobile devices.

Adds methods to enable and disable the ratings widget and adds examples
of how this functionality can be used on the examples page.

Disable the mouse move events on devices with touch events to avoid
requiring a double click on a mobile phones to select a rating.

Updates the README.md and adds a ghpages tasks to publish examples to
GitHub pages.