Reworks Disabled from State into Prop
After talking to Josh, he convinced me that it would be more of the
React way to have disabled as a prop instead of state. If a user of the
widget, wants to disable the widget, they would change the prop instead
of calling a specific method.

Udpates the code to rename the startDisabled prop to disabled and
replaces the disabled state with the disabled prop. Deletes the enable
and disable methods.

Renames onRate to onChange to feel more like normal HTML.

Updates example page to work with changes.